+++ events/local/local.go
@@ -17,7 +17,11 @@ import (
 	. "github.com/mageddo/dns-proxy-server/log"
 )
 
-var confPath string = utils.GetPath(*flags.ConfPath)
+var confPath string = GetConfPath()
+
+func GetConfPath() string {
+	return utils.GetPath(*flags.ConfPath)
+}
 
 func LoadConfiguration(ctx context.Context) (*LocalConfiguration, error){
 
@@ -107,6 +111,8 @@ type LocalConfiguration struct {
 	WebServerPort int `json:"webServerPort"`
 	DnsServerPort int `json:"dnsServerPort"`
 	DefaultDns *bool `json:"defaultDns"`
+	LogLevel string `json:"logLevel"`
+	LogFile string `json:"logFile"`
 }
 
 type EnvVo struct {