+++ resolvconf/dns_server_cleaner.go
@@ -1,4 +1,4 @@
-package conf
+package resolvconf
 
 import "strings"