+++ RELEASE-NOTES.md
@@ -1,3 +1,7 @@
+### 2.4.0
+* Enable/Disable log/set log path using `MG_LOG_FILE` env or `--log-file` command line option or json config
+* Change log level using `MG_LOG_LEVEL` env or `--log-level` command line option or json config
+
 ### 2.3.3
 * Domains wildcard support
 If you register a hostname with `.` at start, then all subdomains will solve to that container/local storage entry